Alarming and stat output

+++ src/shared/ethernet/ethernet_receive_udp.cpp
@@ -158,7 +158,7 @@ void handleIncomingUdpMessage() {
   }
 }
 
-// Report on UDP received message statistics, including messages per second
+// Report on UDP received message statistics, including messages per second (as integer)
 void reportUdpMessageReceiveStats() {
   static unsigned long lastStatsReportTime = 0;
   static unsigned long lastMessageCount = 0;
@@ -166,16 +166,16 @@ void reportUdpMessageReceiveStats() {
   unsigned long currentTime = millis();
   unsigned long elapsedTime = currentTime - lastStatsReportTime;
   unsigned long messagesThisPeriod = receiveUdpTotalMessageCount - lastMessageCount;
-  float messagesPerSecond = 0.0f;
+  unsigned int messagesPerSecond = 0;
 
   if (elapsedTime > 0) {
-    messagesPerSecond = (messagesThisPeriod * 1000.0f) / elapsedTime;
+    messagesPerSecond = (messagesThisPeriod * 1000UL) / elapsedTime;
   }
 
   float malformedPercentage     = (receiveUdpTotalMessageCount > 0) ? (receiveMalformedMessageCount * 100.0f / receiveUdpTotalMessageCount) : 0.0f;
   float outOfSequencePercentage = (receiveUdpTotalMessageCount > 0) ? (receiveOutOfSequenceCount * 100.0f / receiveUdpTotalMessageCount) : 0.0f;
 
-  DEBUG_PERFORMANCE("UDP messages received: %lu, Malformed: %.2f%% (%lu), Out of sequence: %.2f%% (%lu), Rate: %.2f msg/sec",
+  DEBUG_PERFORMANCE("UDP messages received: %lu, Malformed: %.2f%% (%lu), Out of sequence: %.2f%% (%lu), Rate: %u msg/sec",
                     receiveUdpTotalMessageCount, malformedPercentage, receiveMalformedMessageCount,
                     outOfSequencePercentage, receiveOutOfSequenceCount, messagesPerSecond);
